[cookbooks] Convert chef_run resource to new resource style
